Use a randomised Django secret key if one isn't set explicitly (#201)
The Django secret key is no longer hardcoded to an example value, instead
using a randomised value if the recommend env var is not set. The previous
implementation could be insecure if users didn't set an explicit value, and
also encouraged secrets being committed to source rather than the use of
env vars.

The name of the env var has also been renamed from `SECRET_KEY` to
`DJANGO_SECRET_KEY`.